+++ mqmetric/queue.go
@@ -45,6 +45,7 @@ const (
 	ATTR_Q_DEPTH       = "depth"
 	ATTR_Q_SINCE_PUT   = "time_since_put"
 	ATTR_Q_SINCE_GET   = "time_since_get"
+	ATTR_Q_MAX_DEPTH   = "attribute_max_depth"
 )
 
 var QueueStatus StatusSet
@@ -79,11 +80,20 @@ func QueueInitAttributes() {
 	QueueStatus.Attributes[attr] = newStatusAttribute(attr, "Input Handles", ibmmq.MQIA_OPEN_INPUT_COUNT)
 	attr = ATTR_Q_OPPROCS
 	QueueStatus.Attributes[attr] = newStatusAttribute(attr, "Input Handles", ibmmq.MQIA_OPEN_OUTPUT_COUNT)
+
 	// Usually we get the QDepth from published resources, But on z/OS we can get it from the QSTATUS response
-	if platform == ibmmq.MQPL_ZOS {
+	if !usePublications {
 		attr = ATTR_Q_DEPTH
 		QueueStatus.Attributes[attr] = newStatusAttribute(attr, "Queue Depth", ibmmq.MQIA_CURRENT_Q_DEPTH)
 	}
+
+	// This is not really a momitoring metric but it enables calculations to be made such as %full for
+	// the queue. It's extracted at startup of the program via INQUIRE_Q and not updated later even if the
+	// queue definition is changed. It's not easy to generate the % value in this program as the CurDepth will
+	// usually - but not always - come from the published resource stats. So we don't have direct access to it.
+	// Recording the MaxDepth allows Prometheus etc to do the calculation regardless of how the CurDepth was obtained.
+	attr = ATTR_Q_MAX_DEPTH
+	QueueStatus.Attributes[attr] = newStatusAttribute(attr, "Queue Max Depth", -1)
 
 	attr = ATTR_Q_QTIME_SHORT
 	QueueStatus.Attributes[attr] = newStatusAttribute(attr, "Queue Time Short", ibmmq.MQIACF_Q_TIME_INDICATOR)
@@ -178,6 +188,66 @@ func collectQueueStatus(pattern string, instanceType int32) error {
 	return err
 }
 
+// Issue the INQUIRE_Q call for wildcarded queue names and
+// extract the required attributes - currently, just the
+// Maximum Queue Depth
+func inquireQueueAttributes(objectPatternsList string) error {
+	var err error
+
+	statusClearReplyQ()
+
+	if objectPatternsList == "" {
+		return err
+	}
+
+	objectPatterns := strings.Split(strings.TrimSpace(objectPatternsList), ",")
+	for i := 0; i < len(objectPatterns) && err == nil; i++ {
+		var buf []byte
+		pattern := strings.TrimSpace(objectPatterns[i])
+		if len(pattern) == 0 {
+			continue
+		}
+
+		putmqmd, pmo, cfh, buf := statusSetCommandHeaders()
+
+		// Can allow all the other fields to default
+		cfh.Command = ibmmq.MQCMD_INQUIRE_Q
+
+		// Add the parameters one at a time into a buffer
+		pcfparm := new(ibmmq.PCFParameter)
+		pcfparm.Type = ibmmq.MQCFT_STRING
+		pcfparm.Parameter = ibmmq.MQCA_Q_NAME
+		pcfparm.String = []string{pattern}
+		cfh.ParameterCount++
+		buf = append(buf, pcfparm.Bytes()...)
+
+		pcfparm = new(ibmmq.PCFParameter)
+		pcfparm.Type = ibmmq.MQCFT_INTEGER_LIST
+		pcfparm.Parameter = ibmmq.MQIACF_Q_ATTRS
+		pcfparm.Int64Value = []int64{int64(ibmmq.MQIA_MAX_Q_DEPTH)}
+		cfh.ParameterCount++
+		buf = append(buf, pcfparm.Bytes()...)
+
+		// Once we know the total number of parameters, put the
+		// CFH header on the front of the buffer.
+		buf = append(cfh.Bytes(), buf...)
+
+		// And now put the command to the queue
+		err = cmdQObj.Put(putmqmd, pmo, buf)
+		if err != nil {
+			return err
+		}
+
+		for allReceived := false; !allReceived; {
+			cfh, buf, allReceived, err = statusGetReply()
+			if buf != nil {
+				parseQAttrData(cfh, buf)
+			}
+		}
+	}
+	return nil
+}
+
 // Given a PCF response message, parse it to extract the desired statistics
 func parseQData(instanceType int32, cfh *ibmmq.MQCFH, buf []byte) string {
 	var elem *ibmmq.PCFParameter
@@ -248,7 +318,66 @@ func parseQData(instanceType int32, cfh *ibmmq.MQCFH, buf []byte) string {
 	QueueStatus.Attributes[ATTR_Q_SINCE_PUT].Values[key] = newStatusValueInt64(statusTimeDiff(now, lastPutDate, lastPutTime))
 	QueueStatus.Attributes[ATTR_Q_SINCE_GET].Values[key] = newStatusValueInt64(statusTimeDiff(now, lastGetDate, lastGetTime))
 
+	if s, ok := qInfoMap[key]; ok {
+		maxDepth := s.MaxDepth
+		QueueStatus.Attributes[ATTR_Q_MAX_DEPTH].Values[key] = newStatusValueInt64(maxDepth)
+	}
 	return key
+}
+
+func parseQAttrData(cfh *ibmmq.MQCFH, buf []byte) {
+	var elem *ibmmq.PCFParameter
+
+	qName := ""
+
+	parmAvail := true
+	bytesRead := 0
+	offset := 0
+	datalen := len(buf)
+	if cfh.ParameterCount == 0 {
+		return
+	}
+
+	// Parse it once to extract the fields that are needed for the map key
+	for parmAvail && cfh.CompCode != ibmmq.MQCC_FAILED {
+		elem, bytesRead = ibmmq.ReadPCFParameter(buf[offset:])
+		offset += bytesRead
+		// Have we now reached the end of the message
+		if offset >= datalen {
+			parmAvail = false
+		}
+
+		// Only one field needed for queues
+		switch elem.Parameter {
+		case ibmmq.MQCA_Q_NAME:
+			qName = strings.TrimSpace(elem.String[0])
+		}
+	}
+
+	// And then re-parse the message so we can store the metrics now knowing the map key
+	parmAvail = true
+	offset = 0
+	for parmAvail && cfh.CompCode != ibmmq.MQCC_FAILED {
+		elem, bytesRead = ibmmq.ReadPCFParameter(buf[offset:])
+		offset += bytesRead
+		// Have we now reached the end of the message
+		if offset >= datalen {
+			parmAvail = false
+		}
+
+		switch elem.Parameter {
+		case ibmmq.MQIA_MAX_Q_DEPTH:
+			v := elem.Int64Value[0]
+			if v > 0 {
+				if qInfo, ok := qInfoMap[qName]; ok {
+					qInfo.MaxDepth = v
+				}
+			}
+		}
+
+	}
+
+	return
 }
 
 // Return a standardised value.